Assigment2/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
Django==5.2
djangorestframework==3.16.0
fonttools==4.57.0
fpdf2==2.8.2
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
validators==0.34.0
